refactor(tools): separate sha256 functions from tools.utils
