+++ kong-3.6.0-0.rockspec
@@ -167,6 +167,7 @@ build = {
     ["kong.tools.string"] = "kong/tools/string.lua",
     ["kong.tools.table"] = "kong/tools/table.lua",
     ["kong.tools.yield"] = "kong/tools/yield.lua",
+    ["kong.tools.sha256"] = "kong/tools/sha256.lua",
 
     ["kong.runloop.handler"] = "kong/runloop/handler.lua",
     ["kong.runloop.events"] = "kong/runloop/events.lua",
